I have class now, but please run AI_Test.exe
It's set to simulate 20 games for static vs dynamic and 20 games for
dynamic vs static. It'll output the stuff into game_log_2.txt (as well
as cout if you want to watch) That way at least we can analyze the long
term performance of the AIs. It should take max 40 minutes (so you can
let it run in the background while you do other stuff if you want)